Add vendors (#6)
* Add factory_bot_rails and ffaker for use in specs

* Add simple_form for creating forms

* Initial simple_form install

* Add vendors

This adds vendor crud functionality. The functionality
is tested with a system spec.

To add a vendor you must be logged in. 

The initial files were generated with `rails g scaffold` and
then modified. Running that autogenerates a large number
of files that we don't need. Those files were deleted.

Generating a scaffold creates a form, but I'd like to use
`simple_form`. The autogenerated form was deleted and replaced
with the simple_form version.

See: https://github.com/heartcombo/simple_form

For the system spec, I'm using factory_bot to initialize the
models in the spec:

https://github.com/thoughtbot/factory_bot