LAGJH-1047 CRUD for headings.
This adds CRUD for Headings.

This PR has additional database migrations. After
checking out this branch you'll need to run
`docker-compose run --rm hummingbird bundle exec rake db:migrate`
to update the database.

You can test the functionality by logging in and visiting
`/categories/`, `/subcategories/`, `/headings/` and
`/database_headings`.

You'll need to create a a category and a subcategory. After that
you'll be able to create a heading, which is a combination of a
category and subcategory. Then you can associate a database
with a heading.

These are the most basic forms for this functionality. We'll
need to combine some of the forms so that you don't need to
visit the `/database_headings` forms to add a heading to a database.