LAGJH-1047 CRUD for Headings.
This adds CRUD for Headings.

This PR has additional database migrations. After
checking out this branch you'll need to run
docker-compose run --rm hummingbird bundle exec rake db:migrate
to update the database.

You can test the functionality by logging in and visiting
/headings/, /subheadigns and /database_headings.

You'll need to create a heading and a subheading. After that
you'll be able to create a database_heading, which associates
a heading with a database.

These are the most basic forms for this functionality. We'll
need to combine some of the forms so that you don't need to
visit the /database_headings forms to add a heading to a database.